+++ packages/dara-core/changelog.md
@@ -7,6 +7,7 @@ title: Changelog
 -   Fixed an issue where if a `Variable` was created from a `DerivedVariable` and had `persist_value=True` that its value was not unwrapped.
 -   Added "Powered by causaLens" to the sidebar when `powered_by_causaLens` is set to `True` in the configuration.
 -   Added a github link to "Built with Dara" in the sidebar.
+-   Fixed an infinite loop in `useVariable` when using a nested selector inside a Select.
 
 ## 1.10.0